% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.tidytensor.R
\name{as.data.frame.tidytensor}
\alias{as.data.frame.tidytensor}
\title{Convert a tidytensor to a data.frame representation.}
\usage{
\method{as.data.frame}{tidytensor}(tensor, allow_huge = FALSE)
}
\arguments{
\item{allow_huge}{whether to allow the creation of a data.frame with more than 10 million entries}

\item{x}{input to convert to a data.frame}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a data.frame
}
\description{
Given a tidytensor, returns a data.frame, with each rank of the tensor being represented by a column.
Produces an error if the resulting data.frame would have more than 10 million entries; use \code{allow_huge = TRUE} to override.
}
\details{
Note that this produces a row for each value in the tensor, and a column for each rank; data.frames are a much less
efficient representation, but can be useful for e.g. visualization purposes. This method thus produces an error if
the resulting data.frame would have more than 10 million entries; but one can set \code{allow_huge = TRUE} to override.
If dimnames() are set (naming each dimension), then the columns will be factors, rather than integer indices.

If the tidytensor ranks are not named, columns will be named \code{index_1}, \code{index_2}, etc., otherwise they will be
set to ranknames.
Tensor values will be in a column named \code{value}.
}
\examples{
# From an array (representing e.g. 30 26x26 images (30 sets of 26 rows of 26 pixels))
a <- array(rnorm(30 * 26 * 26), dim = c(30, 26, 26))
t <- as.tidytensor(a)
ranknames(t) <- c("sample", "row", "pixel")
df <- as.data.frame(t)
print(head(df))

# Example with named dimensions:
dimnames(t)[[1]] <- paste("sample", 1:30, sep = "_")
dimnames(t)[[2]] <- paste("row", 1:26, sep = "_")
dimnames(t)[[3]] <- paste("pixel", 1:26, sep = "_")
# or with a list:
dimnames(t) <- list(paste("sample", 1:30, sep = "_"),
                    paste("row", 1:26, sep = "_"),
                    paste("pixel", 1:26, sep = "_"))

print(head(as.data.frame(t)))
}
\seealso{
\code{\link{ranknames}}.
}
