% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.tensortree.R
\name{as.list.tensortree}
\alias{as.list.tensortree}
\title{Convert a tensortree into a nested list of tensors.}
\usage{
\method{as.list}{tensortree}(tensor, rank = 1, flatten = FALSE,
  state = NULL)
}
\arguments{
\item{tensor}{the tensortree to convert.}

\item{rank}{an indicator of the rank defining the contained tensors.}

\item{flatten}{whether to return a nested list (\code{FALSE}) or a flattened list (\code{TRUE}).}

\item{state}{an internally used parameter for tracking build state-do not set manually.}

\item{...}{additional arguments passed to methods (unusued).}
}
\value{
a list.
}
\description{
Convert a tensortree into a nested list of tensors, nested down to level specified in \code{rank}.
If \code{flatten = TRUE}, returns a flattens the structure to a list of tensors (not nested).
}
\details{
The \code{state} parameter is for internal use, and needn't be set during normal usage.
}
\examples{
# Three tensortrees of the same shape
t3 <- as.tensortree(array(100 * 1:(3 * 4 * 5), dim = c(3, 4, 5)))
ranknames(t1) <- c("sample", "row", "col")
l1 <- as.list(t1)
str(l1)
}
\seealso{
\code{\link{as.data.frame.tensortree}}
}
