% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ranks.R
\name{combine_ranks}
\alias{combine_ranks}
\title{Combine multiple ranks of a tensor into a single rank}
\usage{
combine_ranks(x, ..., new_rank_name = NULL)
}
\arguments{
\item{x}{the tidytensor to combine ranks for.}

\item{...}{ranknames or integers to combine (quoted or unquoted).}

\item{.dots}{character or integer vector of ranknames.}
}
\value{
a new tidytensor.
}
\description{
Combine multiple ranks of a tensor into a single rank, for example for use in data augmentation.
}
\details{
If all ranks being combined have dimension names, the dimension names of the newly produced rank will be combinations of those specified.

It is only possible to combine consecutive ranks; use \code{permute()} to first organize ranks.
}
\examples{
# shape [5, 20, 26, 26] for 5 batches of 20 26x26 "images"
t <- as.tidytensor(array(rnorm(5 * 20 * 26 * 26), dim = c(5, 20, 26, 26)))
ranknames(t) <- c("batch", "image", "row", "col")

# given an image tidytensor (26x26), return a set of replicates with noise added
make_noisy_images <- function(t2) {
  res <- bind(t2,
              t2 + rnorm(length(t2)),
              t2 + rnorm(length(t2)),
              t2 + rnorm(length(t2)), new_rank_name = "replicate")
}

#augment the original data by replacing each image with a set of noisy replicates
t <- tt_apply(t, image, make_noisy_images)
# now t is shape (5, 20, 4, 26, 26) with ranknames (batch, image, replicate, row, col)
# let's set some dimension names
t <- set_dimnames_for_rank(t, image, .dots = 1:20)  # "1", "2", "3", ...
t <- set_dimnames_for_rank(t, replicate, original, rep1, rep2, rep3) # "original", "rep1", "rep2", "rep3"

# to make it compatible with the original shape we combine images and replicates
t2 <- combine_ranks(t, image, replicate)

print(t2)

# since the combined ranks both have demension names, the newly created rank does as well and we can verify contents
# here we see that the second batch, image 3, replicate 2 is indeed the same
print(t[2, "3", "rep2", , ])
print(t2[2, "3_rep2", , ])
}
\seealso{
\code{\link{permute}}, \code{\link{bind}}
}
