% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_encode_tensor.R
\name{fasta_encode_tensor}
\alias{fasta_encode_tensor}
\title{Parse a FASTA file into a tidytensor, using 1-hot encoding for base representation.}
\usage{
fasta_encode_tensor(fasta_file, start = 1, end = NULL, ids = NULL,
  alphabet = "nucleotide", trim_to = NULL, ...)
}
\arguments{
\item{fasta_file}{input filename to convert to a tidytensor.}

\item{start}{starting sequence number to grab.}

\item{end}{ending sequence number to grab (see details).}

\item{ids}{only grab sequences with these sequence IDs.}

\item{alphabet}{either "nucleotide", "protein", or a named list that specifies the encoding, per base.}

\item{trim_to}{integer length to trim all sequences to if they are longer; don't attempt trimming if NULL.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a new tidytensor.
}
\description{
Parses a FASTA-formatted file, returning a 1-hot (or multi-hot if IUPAC codes are used) encoded tensor of rank 3, with ranknames "Sequence", "Base", and "Channel". Allows the user
to specify that start and end indices for sequences to grab (e.g. grab the 5th to the 20th sequences), or sequence IDs. The default "nucleotide" encoding
accepts lower and upper-case A, C, G, and T and degenerate IUPAC codes. The "protein" alphabet results in length-21 one-hot
encoding (20 amino acid letters, plus stop codon '*'; also, the uknown "X" is encoded as an all-zero vector). Alternatively,
one can set alphabet to a named list for endoding; e.g. `alphabet <- list("A" = c(1, 0), "C" = c(0, 1), "G" = c(0, 0), "T" = c(1, 1)`.
}
\details{
If \code{seqnames} is given, only these are grabbed from the FASTA file for inclusion in the tensor (a warning will be produced
if any requested sequence IDs are not present in the file). If \code{seqnames} isn't given, \code{start} and \code{end} are used, defaulting
to grabbing all sequences in the file. (Warning: the resulting tensor and processing time may be large.)

The \code{alphabet} parameter specifies how each letter in sequences should be encoded, as a named list.
}
\examples{
t <- fasta_encode_tensor("seqs.fasta")
print(t, bottom = "2d", end_n = c(6, 4))

}
\seealso{
\code{\link{ranknames}}.
}
