% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_encode_tensor.R
\name{fasta_train_batch}
\alias{fasta_train_batch}
\title{Parse entries from a FASTA file into a tensortree with target data to predict.}
\usage{
fasta_train_batch(fasta_file, ids, targets,
  keras_categorical_targets = FALSE, keras_format = TRUE,
  alphabet = list(A = c(1, 0, 0, 0), C = c(0, 1, 0, 0), G = c(0, 0, 1,
  0), T = c(0, 0, 0, 1), N = c(0, 0, 0, 0)), ...)
}
\arguments{
\item{fasta_file}{input filename to convert to a tensortree.}

\item{ids}{IDs of sequences to include in the tensor.}

\item{targets}{training targets for ids.}

\item{keras_categorical_targets}{convert targets with \code{keras::to_categorical}.}

\item{keras_format}{format the tensor using Keras/Python-style ordering (row-major), otherwise use default array ordering.}

\item{alphabet}{a list that specifies the one-hot encoding, per base.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a new tensortree.
}
\description{
Given a FASTA filename, a list of sequence IDs, and a list of target values, generates a list with two elements: the first
being the tensor encoding the relavent sequences, and the second being a tensor encoding the targets. If the desired targets are categorical,
\code{keras_categorical_targets = TRUE} uses Keras' \code{to_categorical} function to encode them as one-hot vectors as well.
}
\details{
For options \code{fasta_file}, \code{keras_format}, and \code{alphabet}, see \code{\link{fasta_encode_tensor}}. The \code{ids} and
\code{targets} vectors must be identical in length; if any \code{ids} are specified that are not contained in the FASTA file a warning
will be produced, and the corresponding targets will be removed from the returned targets tensor.

If \code{keras_categorical_targets = TRUE}, the targets are first integer-encoded with integers \code{1} through \code{length(unique(targets))} (using \code{as.integer(factor(targets))}, so generally
integer targets will be assigned in increasing target order), since \code{keras::to_categorical} can only accept integer labels. \code{dimnames()} of
of the target tensor will be set to preserve the original target labels. If \code{targets} is a factor, factor ordering will be used in target assignment.
}
\examples{
t <- fasta_train_batch("seqs.fasta", c("GL15X3", "AGP4", "CATB"), c(1.2, -4.5, 6.7))
print(t, bottom = "2d", end_n = c(6, 4))

t <- fasta_train_batch("seqs.fasta", c("GL15X3", "AGP4", "CATB"), c("positive", "negative", "positive"), keras_categorical_targets = TRUE)
print(t, bottom = "2d", end_n = c(6, 4))
}
\seealso{
\code{\link{fasta_encode_tensor}}.
}
