% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_encode_tensor.R
\name{fastas_to_targets_df}
\alias{fastas_to_targets_df}
\title{Generate a dataframe containing IDs and file-based target labels from a list of fasta-encoded filenames.}
\usage{
fastas_to_targets_df(fasta_files, directory = NULL)
}
\arguments{
\item{fasta_files}{vector of fasta file names.}

\item{directory}{directory to look for fasta files, in; if NULL, use current working directory.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a data.frame.
}
\description{
Given a vector of FASTA filenames, returns a dataframe with columns for "filename", "seqid", and "class", where class is derived from the filename.
Useful in conjunction with \code{\link{flow_sequences_from_fasta_df}}.
#'
}
\examples{
df <- fastas_to_targets_df(c("seqs1.fasta", "seqs2.fasta"))

gen <- flow_sequences_from_fasta_df(df)
}
\seealso{
\code{\link{flow_sequences_from_fasta_df}}.
}
