% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_encode_tensor.R
\name{flow_sequences_from_fasta}
\alias{flow_sequences_from_fasta}
\title{Returns a Keras-style generator for training from sequence data stored in a FASTA file in batches.}
\usage{
flow_sequences_from_fasta(fasta_file, ids, targets,
  keras_categorical_targets = F, batch_size = 20, shuffle = FALSE,
  seed = NULL, ...)
}
\arguments{
\item{fasta_file}{input filename to generate training batches from.}

\item{ids}{IDs of sequences to include in the tensor.}

\item{targets}{training targets for ids.}

\item{keras_categorical_targets}{convert targets with \code{keras::to_categorical}.}

\item{batch_size}{returns training batches of this size.}

\item{shuffle}{should the ids be shuffled first, or batches returned in the order specified?}

\item{seed}{use this random seed if shuffling.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a generator function, generating training batches with each call.
}
\description{
Given a FASTA filename, a list of sequence IDs, and a list of target values, generates a list with two elements: the first
being the tensor encoding the relavent sequences, and the second being a tensor encoding the targets. If the desired targets are categorical,
\code{keras_categorical_targets = TRUE} uses Keras' \code{to_categorical} function to encode them as one-hot vectors as well.
}
\details{
For options \code{fasta_file}, \code{keras_format}, and \code{alphabet}, see \code{\link{fasta_encode_tensor}}. The \code{ids} and
\code{targets} vectors must be identical in length; if any \code{ids} are specified that are not contained in the FASTA file a warning
will be produced, and the corresponding targets will be removed from the returned targets tensor.

If \code{keras_categorical_targets = TRUE}, the targets are first integer-encoded with integers \code{1} through \code{length(unique(targets))} (using \code{as.integer(factor(targets))}, so generally
integer targets will be assigned in increasing target order), since \code{keras::to_categorical} can only accept integer labels. \code{dimnames()} of
of the target tensor will be set to preserve the original target labels.
}
\examples{
gen <- flow_sequences_from_fasta("seqs.fasta", c("GL15X3", "AGP4", "CATB"), c(1.2, -4.5, 6.7))
batch <- gen()
print(batch[[1]])
print(batch[[2]])
}
\seealso{
\code{\link{fasta_encode_tensor}}, \code{\link{fasta_train_batch}}.
}
