% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{partition.tidytensor}
\alias{partition.tidytensor}
\title{Partition a tidytensor into a list of smaller tidytensors of the same rank}
\usage{
\method{partition}{tidytensor}(x, sizes = c(0.5, 0.5))
}
\arguments{
\item{x}{the tidytensor to apply over.}

\item{sizes}{relative sizes of partitions}

\item{...}{unused}
}
\value{
a list of tidytensors.
}
\description{
Partitions a tensor into pieces of sizes relative to \code{sizes}; e.g. a
tensor with shape (24, 50, 50, 3) partitioned with \code{partition(sizes = c(0.5, 0.5))}
results in a list of two tensors of shape (12, 50, 50, 3).

TODO: Ranknames are respected for both inputs and return values.
}
\details{
Entries in \code{sizes} are treated as relative, so \code{sizes = c(2, 1, 1)}
is equivalent to \code{sizes = c(0.5, 0.25, 0.25)}. Non-integer parition boundaries are
rounded down, and this may result in entries with shape (0, ...), but only when
the size of the first rank is smaller than the number of partitions requested.
}
\examples{
# shape [100, 26, 26]
t <- as.tidytensor(array(rnorm(100 * 26 * 26), dim = c(100, 26, 26)))
ranknames(t) <- c("sample", "row", "col")
print(t)

partitions <- partition(t, c(0.2, 0.8))
print(partitions)

}
\seealso{
\code{\link{index}}, \code{\link{c.tidytensor}}, \code{\link{permute.tidytensor}}
}
