% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tensortree.R
\name{print.tensortree}
\alias{print.tensortree}
\title{Print a tensortree.}
\usage{
\method{print}{tensortree}(x, max_per_level = 2, signif_digits = 4,
  end_n = c(6, 6, 3), show_names = FALSE, bottom = "1d", ...)
}
\arguments{
\item{x}{a tensortree to summarize.}

\item{max_per_level}{only show this many sub-tensors per level.}

\item{signif_digits}{number of significant digits to print for numeric tensors.}

\item{end_n}{limit the base-level prints to include this many dimensions of each rank.}

\item{show_names}{show the dimension names, if present, or dimension indices if not in base-level prints.}

\item{bottom}{either "1d", "2d", or "3d" - specifies whether the inner-most tensors should be represented as rank 1, 2, or 3.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\description{
Prints a summary of a tensortree as a nested hierarchy of tensors of lower rank.
}
\details{

}
\seealso{
\code{\link{print.tensortree}}
}
