% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranknames_various.R
\name{ranknames<-}
\alias{ranknames<-}
\title{Assign ranknames to a tensortree.}
\usage{
ranknames(x, ...) <- value
}
\arguments{
\item{x}{input tensortree to set ranknames on.}

\item{...}{additional arguments to be passed to or from methods (ignored).}

\item{value}{what to store in ranknames(x).}
}
\description{
A tensor tree t may have ranknames(t); this is a character vector of the same length as dim(t)
for future use. Note that ranknames(t) is independent of names(t) or dimnames(t); we are not naming elements,
or the dimension names for each rank, but rank names themselves.
Like names() and dimnames(), unset ranknames() are NULL.
}
\details{
Ranknames for a tensortree t are stored as the names() attribute of dimnames(t). If dimnames(t) happens
to be null, before setting ranknames() we create valid dimnames() filled with NA values. The tensortree
package also provides a specialized dimnames() which preserves ranknames when setting dimnames().
}
\examples{
t <- as.tensortree(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
ranknames(t) <- c("sample", "row", "col")
summary(t)

# If given a single rankname, will repeat the rankname for all ranks
t <- as.tensortree(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
ranknames(t) <- "generic"
summary(t)

# works like names():
t <- as.tensortree(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
ranknames(t) <- c("sample", "row", "col")
print(ranknames(t))
ranknames(t)[3] <- "pixel"
summary(t)
}
\seealso{
\code{\link{set_ranknames}}, \code{\link{dimnames<-}}
}
