% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranknames_various.R
\name{ranknames}
\alias{ranknames}
\title{Get ranknames of a tidytensor.}
\usage{
ranknames(x, ...)
}
\arguments{
\item{x}{input tidytensor to get ranknames for.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
character vector of the same length as dim(x), or NULL if unset.
}
\description{
A tidytensor t may have ranknames(t); this is a character vector of the same length as dim(t)
for future use. Note that ranknames(t) is independent of names(t) or dimnames(t); we are not naming elements,
or the dimension names for each rank, but rank names themselves.
Like names() and dimnames(), unset ranknames() are NULL.
}
\details{
Ranknames for a tidytensor t are stored as the names() attribute of dimnames(t). If dimnames(t) happens
to be null, before setting ranknames() we create valid dimnames() filled with NA values. The tidytensor
package also provides a specialized dimnames() which preserves ranknames when setting dimnames().
}
\examples{
t <- as.tidytensor(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
ranknames(t) <- c("sample", "row", "col")
print(ranknames(t))

}
\seealso{
\code{\link{set_ranknames}}, \code{\link{ranknames<-}}
}
