% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta_tensortree.R
\name{read_fasta_tensortree}
\alias{read_fasta_tensortree}
\title{Parse a FASTA file into a keras-compatible tensor.}
\usage{
read_fasta_tensortree(fasta_file, ...)
}
\arguments{
\item{fasta_file}{input filename to convert to a tensortree.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a new tensortree.
}
\description{
Given a filename for a FASTA-formatted file, returns a keras-compatible tensor
(multidimensional array) that is also a tensortree. The shape is [number of sequences, number of bases per sequence, length],
and ranknames ("seq", "base", "channel") are set.
}
\details{
Because the inputs are parsed into a multidimensional array, each sequence in the
file must be the same length. Currently only supports sequences consisting of A, C, T, and G letters.
}
\examples{
t <- read_fasta_tensortree("seqs.fasta")
summary(t)

}
\seealso{
\code{\link{ranknames}}.
}
