% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranknames_various.R
\name{set_ranknames}
\alias{set_ranknames}
\title{Assign ranknames to a tidytensor via a standard function call.}
\usage{
set_ranknames(x, ..., .dots = NULL)
}
\arguments{
\item{x}{input tidytensor to set ranknames on.}

\item{...}{new ranknames to assign (quoted or unquoted).}

\item{.dots}{character vector of new ranknames to assign.}
}
\value{
a tidytensor with ranknames set.
}
\description{
A tidytensor t may have ranknames(t); this is a character vector of the same length as dim(t)
for future use. Note that ranknames(t) is independent of names(t) or dimnames(t); we are not naming elements,
or the dimension names for each rank, but rank names themselves.
Like names() and dimnames(), unset ranknames() are NULL.
}
\details{
Ranknames for a tidytensor t are stored as the names() attribute of dimnames(t). If dimnames(t) happens
to be null, before setting ranknames() we create valid dimnames() filled with NA values. The tidytensor
package also provides a specialized dimnames() which preserves ranknames when setting dimnames().
}
\examples{
t <- as.tidytensor(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
t <- set_ranknames(t, sample, row, col)
t <- set_ranknames(t, .dots = c("sample", "row", "col"))
print(t)
}
\seealso{
\code{\link{ranknames<-}}
}
