% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.tidytensor.R
\name{stitch}
\alias{stitch}
\title{Concatenate two or more tidytensors to create a new one with the same rank}
\usage{
stitch(...)
}
\arguments{
\item{...}{a number of tidytensors of the same shape, or a single list of them.}
}
\value{
a new tidytensor.
}
\description{
Given multiple tidytensors of the same shape except the first rank, concatenates them together to create a tidytensor
of the same shape, but larger in the first. For example, c(x, y, z) where x and have shape [2, 3, 5] and z has shape [10, 3, 5] returns a new tidytensor
of shape [14, 3, 5].
}
\details{
All input tidytensors must have the same shape except for the first rank. If the input ranknames
conflict, only those of the first input tidytensor will be used, and a warning will be generated.
}
\examples{
# Three tidytensors of the same shape
t1 <- as.tidytensor(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
t2 <- as.tidytensor(array(10 * 1:(3 * 4 * 5), dim = c(3, 4, 5)))
t3 <- as.tidytensor(array(100 * 1:(3 * 4 * 5), dim = c(3, 4, 5)))
ranknames(t1) <- c("sample", "row", "col")
ranknames(t2) <- c("sample", "row", "col")
ranknames(t3) <- c("sample", "row", "col")
t4 <- stitch(t1, t2, t3)
print(t4)

list_example <- list(t1, t2, t3)
t5 <- stitch(list_example)
print(t5)
}
\seealso{
\code{\link{bind}}
}
