% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.tidytensor}
\alias{subset.tidytensor}
\title{Subset dimensions of a tidytensor}
\usage{
\method{subset}{tidytensor}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{the tidytensor to apply over.}

\item{...}{named or unnamed parameters specifying subsetting criteria (see examples)}

\item{drop}{whether to drop ranks with size 1 (similar to \code{x[..., drop = TRUE]})}
}
\value{
a tidytensor
}
\description{
A functional form of e.g. \code{tensor[1:10, 3, ]}, supporting selecting by ranknames, usage with %>%, and
indexing when the rank is unknown.
}
\details{
Subsetting a tidytensor with \code{subset()} as opposed to \code{[]} allows for subsetting even when the number of ranks of the input is unknown; see examples.
}
\examples{
# shape [100, 26, 26, 3]
t <- as.tidytensor(array(rnorm(100 * 26 * 26 * 3), dim = c(100, 26, 26, 3)))
ranknames(t) <- c("sample", "row", "col", "channel")
t <- set_dimnames_for_rank(t, channel, R, G, B)
print(t)

t2 <- subset(t, row = 1:10, sample = 27, drop = FALSE)
print(t2)

# same thing, but without named ranks (not a good idea to mixes named
# subsetting and non-named subsetting)
t2 <- subset(t, 27, 1:10, drop = FALSE)
print(t2)

# equiv of t3[1:20, , , c("G", "R", "B")] # since the last rank has dimnames()
# note the re-ordering of channel levels
t3 <- subset(t, sample = 1:20, channel = c("G", "R", "B"), drop = FALSE)
print(t3)

}
\seealso{
\code{\link{shuffle}}, \code{\link{permute}}
}
