% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.tidytensor.R
\name{tt}
\alias{tt}
\title{Convert a vector, matrix, or array to a tidytensor type.}
\usage{
tt(x, ...)
}
\arguments{
\item{x}{input to convert to a tidytensor.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a new tidytensor.
}
\description{
\code{tt()} is a convenience shorthand for \code{as.tidytensor()}. Given a vector, matrix, or array, returns a tidytensor.
If given a vector, converts to a 1-d array supporting \code{dim()}, matrices are left as matrices,
and in all cases the class 'tidytensor' is added.
}
\details{
Matrices are synonymous with 2-d arrays, so these are left as is. Vectors are converted
to 1-d arrays so that they can support \code{dim()}.
}
\examples{
# From an array (representing e.g. 30 26x26 images (30 sets of 26 rows of 26 pixels))
a <- array(rnorm(30 * 26 * 26), dim = c(30, 26, 26))
t <- tt(a)
ranknames(t) <- c("sample", "row", "pixel")
print(t)

# From a matrix (representing e.g. a 26x26 image (26 rows of 26 pixels)) using \%>\%
t <- matrix(rnorm(26 * 26), nrow = 26, ncol = 26) \%>\% tt()
ranknames(t) <- c("row", "pixel")
print(t)

# From a vector (representing e.g. 26 pixel values)
v <- rnorm(26)
t <- tt(rnorm(26))
ranknames(t) <- c("pixel")
print(t)
}
\seealso{
\code{\link{print.tidytensor}}, \code{\link{ranknames}}.
}
