% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_encode_tensor.R
\name{fasta_encode_tensor}
\alias{fasta_encode_tensor}
\title{Parse a FASTA file into a tensortree, using 1-hot encoding for base representation.}
\usage{
fasta_encode_tensor(fasta_file, start = 1, end = NULL, ids = NULL,
  keras_format = TRUE, alphabet = list(A = c(1, 0, 0, 0), C = c(0, 1,
  0, 0), G = c(0, 0, 1, 0), T = c(0, 0, 0, 1), N = c(0, 0, 0, 0)), ...)
}
\arguments{
\item{fasta_file}{input filename to convert to a tensortree.}

\item{start}{starting sequence number to grab.}

\item{end}{ending sequence number to grab (see details).}

\item{ids}{only grab sequences with these sequence IDs.}

\item{alphabet}{a list that specifies the one-hot encoding, per base.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a new tensortree.
}
\description{
Parses a FASTA-formatted file, returning a 1-hot encoded tensor of rank 3, with ranknames "Sequence", "Base", and "Channel". Allows the user
to specify that start and end indices for sequences to grab (e.g. grab the 5th to the 20th sequences), or sequence IDs. The default one-hot encoding
accepts letters "A", "C", "G", "T", and "N" only (with "N" encoded with \code{0, 0, 0, 0}).
}
\details{
If \code{seqnames} is given, only these are grabbed from the FASTA file for inclusion in the tensor (a warning will be produced
if any requested sequence IDs are not present in the file). If \code{seqnames} isn't given, \code{start} and \end{end} are used, defaulting
to grabbing all sequences in the file. (Warning: the resulting tensor and processing time may be large.)

The \code{alphabet} parameter specifies how each letter in sequences should be encoded, as a named list.
}
\examples{
t <- fasta_encode_tensor("seqs.fasta")
print(t, bottom = "2d", end_n = c(6, 4))

}
\seealso{
\code{\link{ranknames}}.
}
