% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.tensortree.R
\name{permute}
\alias{permute}
\title{Permute the ranks of a tensor}
\usage{
permute(tensor, ...)
}
\arguments{
\item{...}{ranknames or integers to permute by (quoted or unquoted).}

\item{x}{the tensortree permute.}

\item{.dots}{character or integer vector to permute by.}
}
\value{
a new tensortree.
}
\description{
Permute the ranks of a tensor, for example to convert between "channels first" and "channels last" representations.

Ranknames are respected for both inputs and return values.
}
\details{
The \code{rank} parameter may be an integer numeric vector (for permuting by index), or character vector (for permuting by rankname).
}
\examples{
# shape [20, 26, 26]
t <- as.tensortree(array(rnorm(20 * 26 * 26), dim = c(20, 26, 26)))
ranknames(t) <- c("sample", "row", "col")
print(t)

t2 <- permute(t, c(3, 1, 2))
t3 <- permute(t, c("col", "sample", "row"))

}
\seealso{
\code{\link{index}}, \code{\link{c.tensortree}}
}
