% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tidytensor.R
\name{print.tidytensor}
\alias{print.tidytensor}
\title{Print a tidytensor.}
\usage{
\method{print}{tidytensor}(
  x,
  show_names = FALSE,
  max_per_level = 1,
  base_rank = NULL,
  max_rows = 6,
  max_cols = 6,
  max_depth = 3,
  signif_digits = 3,
  indent = 0,
  ...
)
}
\arguments{
\item{x}{a tidytensor to summarize.}

\item{show_names}{show the dimension names, if present, or dimension indices if not in base-level prints.}

\item{max_per_level}{only show this many sub-tensors per level.}

\item{base_rank}{either NULL, 1, 2, or 3 - specifies whether the inner/bottom-most tensors should be represented as rank 1, 2, or 3 in a grid (NULL for autodetect based on tensor shape, see details).}

\item{max_rows}{limit the base-level prints to include this many rows (also applies to 1d prints).}

\item{max_cols}{limit the base-level prints to include this many columns.}

\item{max_depth}{in 3d representation, limit the base-level prints to include this many entries of the last rank.}

\item{signif_digits}{number of significant digits to print for numeric tensors.}

\item{indent}{indent the printout by this much (used internally).}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\description{
Prints a summary of a tidytensor as a nested hierarchy of tensors of lower rank.
}
\details{
The \code{base_rank} argument specifies whether the lowest ranks of the tensor (displayed as a grid) should be shown as rank 2 tensors, rank 3 tensors, or rank 1 tensors; the default of \code{NULL} will
select 3 if the last rank is of size 3 or 1 (assuming an image and a "channels-last" convention), 2 if the 3rd-to-last rank is length 3 or 1 (assuming an image
and a "channels-first" convention) or if there are only two ranks or if the last two ranks are equal (assuming an image channel of some kind), and otherwise will default to 1.

\code{max_per_level} indicates how many replicates
}
\examples{
t <- as.tidytensor(array(1:(2 * 3 * 4 * 5), dim = c(2, 3, 4, 5)))
ranknames(t) <- c("samples", "batches", "rows", "cols")
print(t, base_rank = 2)

t <- as.tidytensor(array(1:(2 * 3 * 40 * 50 * 3), dim = c(2, 3, 40, 50, 3)))
ranknames(t) <- c("sample", "batch", "row", "pixel", "channel")
print(t, max_rows = 6, max_cols = 6, max_depth = 3, show_names = TRUE, base_rank = 3)

}
\seealso{
\code{\link{print.tidytensor}}
}
