% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranknames_various.R
\name{set_dimnames_for_rank}
\alias{set_dimnames_for_rank}
\title{Set dimnames() via a standard function call, for a particular rank.}
\usage{
set_dimnames_for_rank(x, rank, ..., .dots = NULL)
}
\arguments{
\item{x}{input tidytensor to set dimnames on.}

\item{rank}{rank to set the dimnames on.}

\item{...}{dimnames to assign (quoted or unquoted).}

\item{.dots}{character vector of dimnames to assign (quoted or unquoted).}
}
\value{
a tidytensor with dimnames set.
}
\description{
Sets the dimensions names for a particular rank, without requiring dimnames for the other ranks.
}
\details{
If all dimnames are unset, they will be set to NA for the other ranks, otherwise they will be left alone.
}
\examples{
t <- as.tidytensor(array(1:(3 * 2), dim = c(3, 2)))
t <- set_dimnames_for_rank(t, 2, valset1, valset2)
t <- set_dimnames_for_rank(t, 2, .dots = c("valset1", "valset2"))
print(t)
}
\seealso{
\code{\link{ranknames<-}}, \code{\link{dimnames}}, \code{\link{set_dimnames}}
}
