% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranknames_various.R
\name{set_ranknames}
\alias{set_ranknames}
\title{Assign ranknames to a tensortree via a standard function call.}
\usage{
set_ranknames(x, newnames, ...)
}
\arguments{
\item{x}{input tensortree to set ranknames on.}

\item{newnames}{character vector of new ranknames to assign.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a tensortree with ranknames set.
}
\description{
A tensor tree t may have ranknames(t); this is a character vector of the same length as dim(t)
for future use. Note that ranknames(t) is independent of names(t) or dimnames(t); we are not naming elements,
or the dimension names for each rank, but rank names themselves.
Like names() and dimnames(), unset ranknames() are NULL.
}
\details{
Ranknames for a tensortree t are stored as the names() attribute of dimnames(t). If dimnames(t) happens
to be null, before setting ranknames() we create valid dimnames() filled with NA values. The tensortree
package also provides a specialized dimnames() which preserves ranknames when setting dimnames().
}
\examples{
t <- as.tensortree(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
t <- set_ranknames(t, c("sample", "row", "col"))
print(t)

# If given a single rankname, will repeat the rankname for all ranks
t <- as.tensortree(array(1:(3 * 4 * 5), dim = c(3, 4, 5)))
t <- set_ranknames(t, "generic")
print(t)

}
\seealso{
\code{\link{ranknames<-}}
}
